+++ app/android/app/build.gradle
@@ -82,8 +82,8 @@ android {
         applicationId "de.davelee.trams"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
-        versionCode 14
-        versionName "0.1.4"
+        versionCode 15
+        versionName "0.1.5"
     }
     signingConfigs {
         debug {
